DTLS/OpenSSL: Added support for PSK
This commit represents the first working DTLS interaction
with another stack: I can now interact with the Ikea Tradfri
Hub using LibNyoci.

The LibNyoci OpenSSL DTLS platform code is somewhat of a mess,
but luckily the mess is isolated to that file. I hope to clean
it up over the next few weeks, and perhaps add support for other
TLS stacks.